requirements.txt
anthropic
pyyaml
python-dotenv
requests
beautifulsoup4
schedule
secure-smtplib
jinja2